core-service : category service and cache

+++ Back-End/core-service/src/main/java/com/aperture/community/core/module/CmsCategoryEntity.java
@@ -47,7 +47,7 @@ public class CmsCategoryEntity implements Serializable {
     @TableField("parent_cid")
 
     @ApiModelProperty(value = "父分类id,顶层默认为0")
-    private Integer parentCid;
+    private Long parentCid;
     /**
      * 分类层级
      */
